Returned the peek and sent the delimeter to the common
