+++ sveshnikov.michail/T2/data-struct.cpp
@@ -1,14 +1,10 @@
 #include "data-struct.hpp"
 #include <iostream>
 #include <stream-guard.hpp>
+#include <delimeter-io.hpp>
 
 namespace
 {
-  struct DelimiterIO
-  {
-    char exp;
-  };
-
   struct SllIO
   {
     long long &ref;
@@ -29,7 +25,6 @@ namespace
     std::string &exp;
   };
 
-  std::istream &operator>>(std::istream &in, DelimiterIO &&dest);
   std::istream &operator>>(std::istream &in, StringIO &&dest);
   std::istream &operator>>(std::istream &in, LabelIO &&dest);
 
@@ -54,22 +49,6 @@ namespace
     return in;
   }
 
-  std::istream &operator>>(std::istream &in, DelimiterIO &&dest)
-  {
-    std::istream::sentry sentry(in);
-    if (!sentry)
-    {
-      return in;
-    }
-    char c = '0';
-    in >> c;
-    if (in && (c != dest.exp))
-    {
-      in.setstate(std::ios::failbit);
-    }
-    return in;
-  }
-
   std::istream &operator>>(std::istream &in, StringIO &&dest)
   {
     std::istream::sentry sentry(in);
@@ -77,7 +56,7 @@ namespace
     {
       return in;
     }
-    return std::getline(in >> DelimiterIO{'"'}, dest.ref, '"');
+    return std::getline(in >> sveshnikov::DelimiterIO{'"'}, dest.ref, '"');
   }
 
   std::istream &operator>>(std::istream &in, LabelIO &&dest)
