+++ sveshnikov.michail/T3/commands.cpp
@@ -239,11 +239,10 @@ void sveshnikov::maxseq(is_t &in, os_t &out, const polygon_set_t &shapes)
   in >> poly;
   if (!in)
   {
-    throw std::logic_error("Error: incorrect polygon!");
+    throw std::invalid_argument("Error: incorrect polygon!");
   }
-  std::string remaining_line;
-  std::getline(in, remaining_line);
-  if (!remaining_line.empty() && remaining_line.find_first_not_of(" ") != std::string::npos)
+  in >> std::ws;
+  if (in.peek() != '\n')
   {
     throw std::invalid_argument("Error: extra characters after command!");
   }
@@ -276,9 +275,8 @@ void sveshnikov::rmecho(is_t &in, os_t &out, polygon_set_t &shapes)
   {
     throw std::invalid_argument("Error: incorrect polygon!");
   }
-  std::string remaining_line;
-  std::getline(in, remaining_line);
-  if (!remaining_line.empty() && remaining_line.find_first_not_of(" ") != std::string::npos)
+  in >> std::ws;
+  if (in.peek() != '\n')
   {
     throw std::invalid_argument("Error: extra characters after command!");
   }
