+++ sveshnikov.michail/common/delimeter-io.cpp
@@ -1,0 +1,17 @@
+#include "delimeter-io.hpp"
+
+std::istream &sveshnikov::operator>>(std::istream &in, DelimiterIO &&dest)
+{
+  std::istream::sentry sentry(in);
+  if (!sentry)
+  {
+    return in;
+  }
+  char c = '0';
+  in >> c;
+  if (in && (c != dest.exp))
+  {
+    in.setstate(std::ios::failbit);
+  }
+  return in;
+}
